auth-api/requirements.txt
# bcrypt 必須 pin 在 4.0.x：
#   passlib 1.7.4 依賴 bcrypt.__about__，bcrypt>=4.1 移除了這個屬性會炸
fastapi==0.115.12
uvicorn[standard]==0.34.2
itsdangerous==2.2.0
passlib==1.7.4
bcrypt==4.0.1
pyyaml==6.0.2
python-multipart==0.0.20
